Update Exchange Rate (tkinter).py
The original code made the API call directly within the convert_currency method. Since requests.get is a synchronous function, it would pause the entire GUI thread until a response was received from the server. This could make the application appear to freeze.

By using Clock.schedule_once(self.convert_currency, 0), we are essentially telling Kivy to run the API call in the next available frame, which allows the "Fetching..." message to display and the GUI to remain responsive.

I've added more specific error messages to guide the user. and improved the interface.